Create Powered Addition
O(N) time complexity.
Difficult for me for the first time.
Evaluate the max difference or the max value a number has to be increased to achieve the given situation. Find this in O(N).
Evaluate Log base 2 the max difference. Take floor. Add 1 and print the answer.(2^(x-1) is the increment!!!)